requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
flask-restplus==0.13.0
Flask-Session==0.3.1
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.0.2
pycryptodome==3.9.4
pyrsistent==0.15.6
pytz==2019.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
zipp==0.6.0
